drm/sun4i: sun8i: Workaround for scaler hardware-related issues
[Problem]
Mainline drm/sun4i DE2 and DE3 always had problems with scaling:
Anytime when a user changes scale ratio on-screen glitches appear and in
some cases the whole pipeline can freeze which requires system restart.

Users who can benefit from scale-based animation have to disable
hardware acceleration and use GPU for such cases which affects
performance.

[Investigation]
I noticed that the top part of the screen usually does not have any
glitches while after "ray" passes some vertical line glitches start
to appear. Later-on I discovered that a glitch appeared once a new
commit was sent to the DRM and the vertical line that splits the screen
into top and bottom half depends on the time offset commit was sent.

All this guided me to the set of scaler hardware registers that for some
reason does not wait for VSYNC edge and instead is used by hardware
immediately after being set.

Broken registers can be easily probed using devmem.
To do that first we must prepare some scale-containing commit and keep it
on the screen. Then using devmem tool change register values. In case
something changes on the screen, the register has a double-buffering issue.
If change can be seen only after storing 1 into the mixer.dbuff register,
then the register doesn't have a double buffering issue.

Using the method described above the following broken register set was
discovered:

 - VSU.CTRL.EN

 - UIS.CTRL.EN
 - UIS.OUTSIZE
 - UIS.INSIZE
 - UIS.HSTEP
 - UIS.VSTEP
 - UIS.HPHASE
 - UIS.VPHASE0

The only solution (to be precise - workaround) that I can see as far is
to set these registers during VBLANK.

[Challenges]
1.a. VBLANK interrupt can be temporarily disabled. This situation can
     occur when a lot of serial console dmesg output are pending.
1.b. VBLANK interrupt can be delayed while CPU is handling another
     interrupt.
1.c. VBLANK duration can be smaller than required due to small display
     mode vporch values

2.a. Difference between UIS[ch].OutSize and BLD[z].InSize can cause
     pipeline freeze.
2.b. UIS[ch].OutSize is set once the register is updated while
     BLD[z].InSize is latched on vsync negative edge.

3.a. Race conditions can occur when VSYNC finishes before we set all the
     registers.

[Conclusion]
There is no reliable solution for this problem. But there is a workaround
that can eliminate pipeline freeze cases and can significantly reduce
screen glitches, which is enough to enable scale-based animations for
most of the cases.

[Implementation details]
Challenge 2.a. and 2.b. solution:
1. Instead of setting the scaler registers every frame (which is also
   consumes a lot of GPU cycles), track the changes of the scaling ratio
   and skip any scaler regs setup activity if scaling ratio or phase
   have not changed.

2. Once scaling ratio has changed set the BLD[z].InSize to 1x1 pixels,
   which help to prevent pipeline freeze in case setting-up of broken
   registers will be delayed.
   BLD[z].InSize will be set to correct value once all broken registers
   will be updated.
   This results correct appearence of scaled layer in case we set all
   necessary regs during the VBLANK. Otherwise we will see empty layer
   (1x1 size) during 1 frame.
   On practice it's very rare case on fast SOCs (H6), but chance to
   miss VBLANK is much higher on slow SOC especially when drm mode
   has small VBLANK lines count.

Signed-off-by: Roman Stratiienko <r.stratiienko@gmail.com>