drm_hwcomposer: Split the composition cleanup and display's deactivation
When drm_hwcomposer is used with drm/sun4i driver,
'adb shell stop && adb shell start' sequence causes noise on the
unused planes during the bootanimation.

It looks like some bug in the drivers or DRM frontend itself, which
keeps the plane assigned to the CRTC if the atomic commit contains CRTC
deactivation.

We need more time to investigate it. For now we can split disposal
into 2 atomic commits, which fixes the noise issue.

Signed-off-by: Roman Stratiienko <roman.o.stratiienko@globallogic.com>