drm_hwcomposer: Workaround for screen de-activating causing db845c regression
Since commit d0494d9b8097 ("drm_hwcomposer: Fixes for display
hotplug / headless mode"), which introduced an extra call to
de-active the display on de-init, we've seen regressions on db845c
with the VtsHalGraphicsComposerV2_3TargetTest test set.

This seems to be due to an issue on db845c, where after putting the
lt9611 bridge to sleep, the connection detection stops working even
after the chip is reset.

However, on changing the kernel driver to avoid putting the bridge
to sleep so connect detection works, we start to see SIGSEGV crashes
in drm_hwcomposer, due to the service expecting the hwmodule to
finish registercallback during client detaching within 1 second.
Exceeding this values causes VTS to crash. And msm driver introduces
delay enough to exceed the limit.

So this patch is a workaround to restore previous behavior to avoid
the regression until we can sort out a proper fixes to db845c and
drm_hwcomposer.

Signed-off-by: John Stultz <jstultz@google.com>
Change-Id: Iff5d3cf3f7fa2f1f2acb68a5a629677137e51188