i915: Vertical align video encode buffers to macroblock size
This is the reland of crrev.com/c/4615698. It is reverted because it
causes the camera regression (b/289983468). The camera IPU hardware
cannot handle a frame whose Y plane and UV planes are not contiguous.
Aligning the height by 16 makes the space between the planes.
This CL aligns the height by 8. This will not cause the camera
problem today because all the heights of camera stack exposed as
supported are aligned by 8.

BUG=b:274756117, b:289983468
TEST=PSNR 31.946 => 34.012 in video.EncodeAccelPerf.av1_180p_desktop2 on rex

Change-Id: I96907ea372bae6740559a67f71856a89811c7cc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/minigbm/+/4739135
Reviewed-by: Chia-I Wu <olv@google.com>
Tested-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>