i915.c: Add log for failure to get MCC environment var
Previously, when the environment variable is not found, IOW when the
feature flag does not set the value for the variable,
|enable_intel_media_compression_var| is NULL and minigbm would log about
this. This would create noise in the log, especially when
`drv->compression` is not true or when the feature flag has not landed.

This CL makes change to only log about the failure to get the
environment variables for Media compression support if
`drv->compression == true`.

Change-Id: Ia6d33b08759d4fe94f2bc8de3fb99287748442b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/minigbm/+/4919324
Tested-by: Hang Nguyen <hnt@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@google.com>
Reviewed-by: Nathan Hebert <nhebert@chromium.org>
Reviewed-by: Dominik Behr <dbehr@chromium.org>
Commit-Queue: Hang Nguyen <hnt@chromium.org>